gcf_cleanup/requirements.txt
httpx
google-cloud-storage
